220: fixes #219
