+++ src/main/java/tech/units/indriya/format/SimpleQuantityFormat.java
@@ -29,7 +29,7 @@
  */
 package tech.units.indriya.format;
 
-import static tech.units.indriya.format.CommonFormatter.parseCompound;
+import static tech.units.indriya.format.CommonFormatter.*;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -72,9 +72,9 @@ import tech.units.indriya.quantity.Quantities;
  *         <td><code>m</code>
  *    <tr>
  *         <td><code>~</code>
- *         <td>Mix
+ *         <td>Mixed radix
  *         <td><a href="#text">Text</a>
- *         <td><code>m</code>; <code>cm</code>
+ *         <td><code>1 m</code>; 27 <code>cm</code>
  * </tbody>
  * </table>
  * </blockquote>
@@ -92,9 +92,12 @@ import tech.units.indriya.quantity.Quantities;
  *     number of digits, and shorter numbers are zero-padded to this amount.
  *     For parsing, the number of pattern letters is ignored unless
  *     it's needed to separate two adjacent fields.<br><br></li>
+ *     
+ *<li><strong><a id="radix">Mixed Radix:</a></strong>
+ *     The Mixed radix marker <code>"~"</code> is followed by a character sequence acting as mixed radix delimiter. This character sequence must not contain <code>"~"</code> itself or any numeric values.<br></li>
  * </ul>
  * </p>
- * @version 1.3.1, $Date: 2019-04-12 $
+ * @version 1.4, $Date: 2019-04-14 $
  * @since 2.0
  */
 @SuppressWarnings("rawtypes")
@@ -106,7 +109,7 @@ public class SimpleQuantityFormat extends AbstractQuantityFormat {
 
 	private static final String NUM_PART = "n";
 	private static final String UNIT_PART = "u";
-	private static final String MIXER = "~";
+	private static final String RADIX = "~";
 	
 	/**
 	 * The pattern string of this formatter. This is always a non-localized pattern.
@@ -139,9 +142,9 @@ public class SimpleQuantityFormat extends AbstractQuantityFormat {
 	public SimpleQuantityFormat(String pattern) {
 		this.pattern = pattern;
 		if (pattern != null && !pattern.isEmpty()) {
-		   if (pattern.contains(MIXER)) {
-		       final String singlePattern = pattern.substring(0, pattern.indexOf(MIXER));
-		       mixDelimiter = pattern.substring(pattern.indexOf(MIXER) + 1);
+		   if (pattern.contains(RADIX)) {
+		       final String singlePattern = pattern.substring(0, pattern.indexOf(RADIX));
+		       mixDelimiter = pattern.substring(pattern.indexOf(RADIX) + 1);
 		       delimiter = singlePattern.substring(pattern.indexOf(NUM_PART)+1, pattern.indexOf(UNIT_PART));
 		   } else {
 		       delimiter = pattern.substring(pattern.indexOf(NUM_PART)+1, pattern.indexOf(UNIT_PART));
@@ -195,9 +198,9 @@ public class SimpleQuantityFormat extends AbstractQuantityFormat {
 	@Override
 	public Quantity<?> parse(CharSequence csq, ParsePosition cursor) throws MeasurementParseException {
         if (mixDelimiter != null && !mixDelimiter.equals(delimiter)) {
-            return parseCompound(csq.toString(), NumberFormat.getInstance(), SimpleUnitFormat.getInstance(), delimiter, mixDelimiter, cursor.getIndex());
+            return parseCompoundAsLeading(csq.toString(), NumberFormat.getInstance(), SimpleUnitFormat.getInstance(), delimiter, mixDelimiter, cursor.getIndex());
         } else if (mixDelimiter != null && mixDelimiter.equals(delimiter)) {
-            return parseCompound(csq.toString(), NumberFormat.getInstance(), SimpleUnitFormat.getInstance(), delimiter, cursor.getIndex());
+            return parseCompoundAsLeading(csq.toString(), NumberFormat.getInstance(), SimpleUnitFormat.getInstance(), delimiter, cursor.getIndex());
         }
 	    int startDecimal = cursor.getIndex();
 		while ((startDecimal < csq.length()) && Character.isWhitespace(csq.charAt(startDecimal))) {
@@ -213,20 +216,9 @@ public class SimpleQuantityFormat extends AbstractQuantityFormat {
 		return Quantities.getQuantity(decimal, unit);
 	}
 
-	@SuppressWarnings("unchecked")
 	@Override
 	protected Quantity<?> parse(CharSequence csq, int index) throws MeasurementParseException {
-		int startDecimal = index; // cursor.getIndex();
-		while ((startDecimal < csq.length()) && Character.isWhitespace(csq.charAt(startDecimal))) {
-			startDecimal++;
-		}
-		int endDecimal = startDecimal + 1;
-		while ((endDecimal < csq.length()) && !Character.isWhitespace(csq.charAt(endDecimal))) {
-			endDecimal++;
-		}
-		Double decimal = Double.valueOf(csq.subSequence(startDecimal, endDecimal).toString());
-		Unit unit = SimpleUnitFormat.getInstance().parse(csq, index);
-		return Quantities.getQuantity(decimal, unit);
+		return parse(csq, new ParsePosition(index));
 	}
 
 	@Override
@@ -268,7 +260,7 @@ public class SimpleQuantityFormat extends AbstractQuantityFormat {
 	}
 	
     @Override
-    protected StringBuffer formatComposite(CompoundQuantity<?> comp, StringBuffer dest) {
+    protected StringBuffer formatCompound(CompoundQuantity<?> comp, StringBuffer dest) {
         final StringBuffer sb = new StringBuffer();
         int i = 0;
         for (Quantity<?> q : comp.getQuantities()) {
