+++ src/main/java/tech/units/indriya/internal/function/calc/DefaultNumberSystem.java
@@ -1,8 +1,46 @@
-package tech.units.indriya.internal.calc;
+/*
+ * Units of Measurement Reference Implementation
+ * Copyright (c) 2005-2019, Units of Measurement project.
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions
+ *    and the following disclaimer in the documentation and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of JSR-385, Indriya nor the names of their contributors may be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+ * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package tech.units.indriya.internal.function.calc;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
+import tech.units.indriya.function.Calculus;
+
+/**
+ * {@link NumberSystem} implementation to support Java's built-in {@link Number}s and the
+ * {@link RationalNumber} type.   
+ * 
+ * @author Andi Huber
+ * @since 2.0
+ */
 public class DefaultNumberSystem implements NumberSystem {
 
     @Override
@@ -70,16 +108,16 @@ public class DefaultNumberSystem implements NumberSystem {
             return RationalNumber.of(BigInteger.ONE, integerToBigInteger(number));
         }
         if(number instanceof BigDecimal) {
-            return BigDecimal.ONE.divide(((BigDecimal) number));
+            return BigDecimal.ONE.divide(((BigDecimal) number), Calculus.MATH_CONTEXT);
         }
         if(number instanceof RationalNumber) {
             return ((RationalNumber) number).reciprocal();
         }
         if(number instanceof Double) {
-            return BigDecimal.ONE.divide(BigDecimal.valueOf((double)number));
+            return BigDecimal.ONE.divide(BigDecimal.valueOf((double)number), Calculus.MATH_CONTEXT);
         }
         if(number instanceof Float) {
-            return BigDecimal.ONE.divide(BigDecimal.valueOf(number.doubleValue()));
+            return BigDecimal.ONE.divide(BigDecimal.valueOf(number.doubleValue()), Calculus.MATH_CONTEXT);
         }
         throw unsupportedNumberType(number);
     }
